feat(publish): Adopt functionality from `np`.

+++ checkpoint/dist-src/commands/publish.js
@@ -80,15 +80,16 @@ export class Publish {
         if (runCleanup) {
             steps.push(async (curr, total) => {
                 this.reporter.step(curr, total, 'Cleanup', '✨');
-                await fs.unlink('package-lock.json');
-                await fs.unlink('yarn.lock');
-                await fs.unlink('node_modules');
+                const hasLockFile = await fs.exists(options.yarn ? 'yarn.lock' : 'package-lock.json');
+                if (!hasLockFile) {
+                    await fs.unlink('node_modules');
+                }
                 await fs.unlink('pkg');
                 if (options.yarn) {
-                    return execa('yarn', ['install', '--production=false']);
+                    return execa('yarn', ['install', '--frozen-lockfile', '--production=false']);
                 }
                 else {
-                    return execa('npm', ['install', '--no-production']);
+                    return execa('npm', hasLockFile ? ['ci'] : ['install', '--no-production', '--no-package-lock']);
                 }
             });
         }
@@ -101,7 +102,11 @@ export class Publish {
             this.reporter.step(curr, total, 'Building Package', '✨');
             const oldIsSilent = reporter.isSilent;
             reporter.isSilent = true;
-            const builder = new Build({ out, publish: true, silent: true }, config, reporter);
+            const builder = new Build({
+                out,
+                publish: true,
+                silent: true,
+            }, config, reporter);
             await builder.init(true);
             reporter.isSilent = oldIsSilent;
         });
@@ -144,7 +149,14 @@ export class Publish {
 export async function run(config, reporter, flags, args) {
     await config.loadPackageManifest();
     const options = args.length > 0
-        ? Object.assign({ cleanup: true }, flags, { yarn: hasYarn(), version: args[0] }) : await ui(Object.assign({}, flags, { yarn: hasYarn() }), config.manifest);
+        ? {
+            cleanup: true,
+            // publish: true,
+            ...flags,
+            yarn: hasYarn(),
+            version: args[0],
+        }
+        : await ui({ ...flags, yarn: hasYarn() }, config.manifest);
     if (!options.confirm) {
         return;
     }
